fix(components): update imports to renamed components
Update FollowUsPopover and SharingOptions to import from renamed
ButtonConfirm component.
